refactor: Add customer repository to OrderUsecase constructor
The code changes add the customer repository to the constructor of the OrderUsecase in the route.go file. This change ensures that the OrderUsecase has access to the customer repository for performing operations related to customers.

Note: The commit message has been generated based on the provided code changes and recent commits.